Requirements
TheMathGame:

Your task is to develop a program that will teach youngsters the basic math facts of addition, subtraction, multiplication and division. The program generates random math problems for students to answer. Players get a small reward for correct answers and suffer a small penalty for incorrect ones. User statistics need to be recorded in a text file so that they're loaded back into the program when the player returns to play the game again. In addition, the youngster should be allowed to see how (s)he is doing at any time while (s)he is playing the game. 

 

WARNING:

Note that your main() function should consist of mostly function calls and not be greater than 100 lines in total; however, your entire program will be closer to 500 - 800 lines of code!

 

PROGRAM REQUIREMENTS

Generate simple math fact problems:
Addition (the total must be an integer >= 0)
Subtraction (the difference must be an integer  >= 0)
Multiplication (the product must be an integer >= 0)
Division (the quotient must be an integer >= 0)
Validate user input at every opportunity. If your program crashes because you allow the user to make an invalid entry and you did not write code to handle such potential exceptions, you will NOT pass the midterm!
The program should keep track of the following statistics:
The user’s name
The total correct answers
The total wrong answers
The total earnings ($0.05 is awarded for every correct response and $0.03 is subtracted from every incorrect response)
A separate text file must be created for every user:
Statistics are read from the file at the start of the game (if the user is a returning player).
Statistics are recorded at the end of every game.
The program must be developed using functions so that the main() function consists mostly of function calls. Below is a list of most of the required functions, you may add your own functions if you like:
credits //This function is used to display your name and what the program does
menu // This function is used to display the menu with various options
validateUserResponse // This function is used to validate user input from the menu
validateUserAnswer // This function is used to validate user input and ensure that ONLY numeric answers are entered by the user.
checkUserAnswer // given a math problem, this function is used to check if the answer the user entered is correct or incorrect
updateStats // This function is used to keep a running total of game statistics (in RAM)
displayStats // This function is used to display statistics on screen
retireveStats //  This function is used to retrieve player statistics from external txt file when the game starts, assuming the player is a returning player, else create a text file to store the stats.
saveStats //  This function is used to save player statistics on an external txt file.
You may also want to consider the following four functions: generateAddition, generateSubtraction, generateMultiplication and generateDivision // use these to generate a problem of the appropriate type.
You must use meaningful variable names.
You must comment your code.
You must use variables of the correct type and initialize them with a proper value.
 

GENERAL RESTRICTIONS FOR ALL QUIZZES, MIDTERM AND FINAL EXAM

No infinite loops, examples include:
for(;;)
while(1)
while(true)
do{//code}while(1);
No break statements to exit loops
No labels or go-to statements